+++ cligen_gen.c
@@ -429,6 +429,11 @@ co_copy(cg_obj  *co,
 		fprintf(stderr, "%s: strdup: %s\n", __FUNCTION__, strerror(errno));
 		return -1;
 	    }
+	if (co->co_translate_fn_str)
+	    if ((con->co_translate_fn_str = strdup(co->co_translate_fn_str)) == NULL){
+		fprintf(stderr, "%s: strdup: %s\n", __FUNCTION__, strerror(errno));
+		return -1;
+	    }
 	if (co->co_show)
 	    if ((con->co_show = strdup(co->co_show)) == NULL){
 		fprintf(stderr, "%s: strdup: %s\n", __FUNCTION__, strerror(errno));
@@ -627,7 +632,7 @@ co_eq(cg_obj *co1,
 	    eq = str_cmp(co1->co_expand_fn_str, co2->co_expand_fn_str);
 	    goto done;
 	}
-
+	/* Should we examine co_translate_fn_str? */
 	/* Examine choice: at least one set, and then strcmp */
 	if (co1->co_choice!=NULL || co2->co_choice!=NULL){
 	    eq = str_cmp(co1->co_choice, co2->co_choice);
@@ -824,6 +829,8 @@ co_free(cg_obj *co,
     if (co->co_type == CO_VARIABLE){
 	if (co->co_expand_fn_str)
 	    free(co->co_expand_fn_str);
+	if (co->co_translate_fn_str)
+	    free(co->co_translate_fn_str);
 	if (co->co_show)
 	    free(co->co_show);
 	if (co->co_expand_fn_vec)