+++ cligen_read.c
@@ -607,7 +607,7 @@ cliread_parse (cligen_handle h,
 	    *co_orig = match_obj->co_ref;
 	else
 	    *co_orig = match_obj;
-	if (cvec_match(*co_orig, string, cvv) < 0){
+	if (cvec_match(h, *co_orig, string, cvv) < 0){
 	    retval = CG_ERROR;
 	    goto done;
 	}